requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
boto3==1.40.32
botocore==1.40.32
build==1.3.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
dnspython==2.8.0
executing==2.2.1
fastapi==0.116.2
fonttools==4.60.0
from-root==1.3.0
h11==0.16.0
idna==3.10
imbalanced-learn==0.14.0
imblearn==0.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mypy-boto3-s3==1.40.26
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
pip-tools==7.5.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.1
pyparsing==3.2.4
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
-e git+https://github.com/P-Saroha/End-to-End-Vehicle-Insurance-Project.git@6761051b0b5f483c3021dfd42f516963e56a429b#egg=src
stack-data==0.6.3
starlette==0.48.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
